[editor] Fix compilation on VS

+++ Tools/Editor/Editor/SceneView.cpp
@@ -508,17 +508,17 @@ void svDrawSprites( SceneView* sv, unsigned screenWidth, unsigned screenHeight )
             opacity = distance / lerpDistance;
         }
         
-        const Vec3 screenPoint = camera->GetScreenPoint( goTransform->GetLocalPosition(), screenWidth, screenHeight );
+        const Vec3 screenPoint = camera->GetScreenPoint( goTransform->GetLocalPosition(), (float)screenWidth, (float)screenHeight );
         const Vec3 viewDir = cameraTransform->GetViewDirection();
         const Vec3 lightDir = (goTransform->GetLocalPosition() - cameraTransform->GetLocalPosition()).Normalized();
         const float viewDotLight = Vec3::Dot( viewDir, lightDir ) ;
         
         if (viewDotLight <= 0 &&
-            screenPoint.x > -texWidth && screenPoint.y > -texHeight &&
+            screenPoint.x > -(float)texWidth && screenPoint.y > -(float)texHeight &&
             screenPoint.x < screenWidth && screenPoint.y < screenHeight)
         {
             const float size = screenHeight / distance;
-            ae3d::System::Draw( &sv->lightTex, screenPoint.x, screenPoint.y, texWidth, texHeight, screenWidth, screenHeight, Vec4( 1, 1, 1, 1 ), ae3d::System::BlendMode::Off );
+            ae3d::System::Draw( &sv->lightTex, (int)screenPoint.x, (int)screenPoint.y, texWidth, texHeight, screenWidth, screenHeight, Vec4( 1, 1, 1, 1 ), ae3d::System::BlendMode::Off );
         }
     }
 }
